libvge: preliminary work to enable static build of libvge on Linux
This work will be a two step process. First, make libvge use regular
Cgo calls, thus update tools/codegen to output the Cgo wrapper code.

Step two would be to build libvge as an object archive; i.e. libvge.a.

Finally, the linker would be updated to link against libvge.a, thus
producing a static binary. At this point, we would start to explore
if it would be possible to transition back to not using Cgo on Linux,
but even if Cgo is still used, we would no longer have to update
LD_LIBRARY_PATH just to have the binaries be able to locate libvge.so.